Update Python example (#385)
* Update Python Example (WIP)

* Update root now.json

* Typo
