+++ python/www/index.html
@@ -1,0 +1,27 @@
+<!doctype html>
+<html>
+  <head>
+    <title>A simple deployment with the Now API!</title>
+    <link rel="stylesheet" href="/www/style.css">
+  </head>
+  <body>
+    <h1>Welcome to a Static Site with Data from a Python API</h1>
+    <h2>Deployed with <a href="https://zeit.co/docs" target="_blank" rel="noreferrer noopener">ZEIT Now</a>!</h2>
+    <p><a href="https://github.com/zeit/now-examples/blob/master/python" target="_blank" rel="noreferrer noopener">This project</a> includes three files. A static <code>www/index.html</code> (rewritten with <a href="https://zeit.co/docs/v2/deployments/routes/" target="_blank" rel="noreferrer noopener">Routes</a> for access from the root) file as the homepage, a static <code>www/style.css</code> file for styling, and a <code>api/date.py</code> serverless function that returns the date on invocation. See <a href="/api/date.py"><code>api/date.py</code> for the Date API with Python</a>.</p>
+    <br />
+    <h3>The date according to Python is:</h3>
+    <p class="date js-date">Loading date...</p>
+
+    <script>
+      function main() {
+        return fetch("/api/date.py").then(function (response) {
+          return response.text();
+        }).then(function (date) {
+          document.querySelector(".js-date").textContent = date
+        });
+      }
+
+      main();
+    </script>
+  </body>
+</html>
