+++ python/www/style.css
@@ -1,0 +1,53 @@
+body {
+  font-family: "SF Pro Text","SF Pro Icons","Helvetica Neue","Helvetica","Arial",sans-serif;
+  padding: 20px 20px 60px;
+  max-width: 680px;
+  margin: 0 auto;
+  font-size: 16px;
+  line-height: 1.65;
+  word-break: break-word;
+  font-kerning: auto;
+  font-variant: normal;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  text-rendering: optimizeLegibility;
+  hyphens: auto;
+  text-align: center;
+}
+
+h2, h3, h4 {
+  margin-top: 1.5em;
+}
+
+h1 {
+  margin-top: 70px; 
+  text-align: center; 
+  font-size: 45px; 
+} 
+
+a {
+  cursor: pointer;
+  color: #0076FF;
+  text-decoration: none;
+  transition: all 0.2s ease;
+  border-bottom: 1px solid white;
+}
+
+a:hover {
+  border-bottom: 1px solid #0076FF;
+}
+
+code, pre {
+  font-family: Menlo, Monaco, Lucida Console, Liberation Mono, DejaVu Sans Mono, Bitstream Vera Sans Mono, Courier New, monospace, serif;
+  font-size: .92em;
+  color: #D400FF;
+}
+
+code:before, code:after {
+  content: '`';
+}
+
+p {
+  text-align: center; 
+  font-size: 16px; 
+}
